requirements.txt
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
h5py==3.7.0
idna==3.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==4.6.0.66
Pillow==9.3.0
PyYAML==6.0
requests==2.28.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
